+++ R/getAdmin.R
@@ -44,9 +44,10 @@
 #' \dontrun{
 #' library(amapGeocode)
 #'
-#' # Set the amap_key which is applied from 'AutoNavi' Map Services for amapGeocde globally.
-#' options(amap_key = 'REPLACE THIS BY YOUR KEY')
-#
+#' # Before the `getAdmin()` is executed,
+#' # the token should be set by `option(amap_key = 'key')`
+#' # or set by key argument in `getAdmin()`
+#'
 #' # Get subordinate administrative regions as a tibble
 #' getAdmin('Sichuan Province')
 #' # Get subordinate administrative regions as a XML
@@ -55,8 +56,95 @@
 #' }
 #'
 #' @seealso \code{\link{extractAdmin}}
-
 getAdmin <-
+  function(keywords,
+           key = NULL,
+           subdistrict = NULL,
+           page = NULL,
+           offset = NULL,
+           extensions = NULL,
+           filter = NULL,
+           callback = NULL,
+           output = NULL,
+           to_table = TRUE) {
+    if (length(keywords) == 1) {
+      # if there is one address, use getCoord.individual directly
+      getAdmin.individual(
+        keywords,
+        key = key,
+        subdistrict = subdistrict,
+        page = page,
+        offset = offset,
+        extensions = extensions,
+        filter = filter,
+        callback = callback,
+        output = output,
+        to_table = to_table
+      )
+    } else {
+      # if there is multiple addresses, use getCoord.individual by laapply
+      ls_queries <-
+        purrr::map(
+          keywords,
+          getAdmin.individual,
+          key = key,
+          subdistrict = subdistrict,
+          page = page,
+          offset = offset,
+          extensions = extensions,
+          filter = filter,
+          callback = callback,
+          output = output,
+          to_table = to_table
+        )
+      # here, getAdmin doesn't support bind_rows
+      # because what the `getAdmin.individual` get general is a tibble
+      # `bind_rows` has the potential probability to confuse the dimension of tibbles
+      return(ls_queries)
+    }
+  }
+
+#' Get an individual tibble of Subordinate Administrative Regions from location
+#'
+#' @param keywords Required.\cr
+#' Search keywords. \cr
+#' Rules: Country/Region, Province/State, City, County/District, Town, Country, Road, Number, Room, Building.
+#' @param key Optional.\cr
+#' Amap Key. \cr
+#' Applied from 'AutoNavi' Map API official website\url{https://lbs.amap.com/dev/}
+#' @param subdistrict Optional.\cr
+#' Subordinate Administrative Level.\cr
+#' Display level of subordinate administrative regions. Available value: 0,1,2,3.\cr
+#' `0` do not return subordinate administrative regions.\cr
+#' `1` return first one subordinate administrative regions.\cr
+#' `2` return first two subordinate administrative regions.\cr
+#' `3` return first three subordinate administrative regions.
+#' @param page Optional.\cr
+#' Which page to return.\cr
+#' Each time the outmost layer will return a maximum of 20 records. If the limit is exceeded, please request the next page of records with the page argument.
+#' @param offset Optional.\cr
+#' Maximum records per page.\cr
+#' Maximum value is 20.
+#' @param extensions Optional.\cr
+#' Return results controller.\cr
+#' `base`: does not return the coordinates of the administrative district boundary.\cr
+#' `all`: returns only the boundary value of the current query district, not the boundary value of the child node.
+#' @param filter Optional.\cr
+#' Filter administrative regions.\cr
+#' Filtering by designated administrative divisions, which returns information only for the province/municipality.\cr
+#' It is strongly recommended to fill in this parameter in order to ensure the correct records.
+#' @param callback Optional.\cr
+#' Callback Function. \cr
+#' The value of callback is the customized function. Only available with JSON output.
+#' If you don't understand, it means you don't need it, just like me.
+#' @param output Optional.\cr
+#'  Output Data Structure. \cr
+#' Support JSON and XML. The default value is JSON.
+#' @param to_table Optional.\cr
+#' Transform response content to tibble.\cr
+#' @return
+#' Returns a JSON or XML of results containing detailed subordinate administrative region information. See \url{https://lbs.amap.com/api/webservice/guide/api/district} for more information.
+getAdmin.individual <-
   function(keywords,
            key = NULL,
            subdistrict = NULL,
@@ -127,8 +215,9 @@ getAdmin <-
 #' library(dplyr)
 #' library(amapGeocode)
 #'
-#' # Set the amap_key which is applied from 'AutoNavi' Map Services for amapGeocde globally.
-#' options(amap_key = 'REPLACE THIS BY YOUR KEY')
+#' # Before the `getAdmin()` is executed,
+#' # the token should be set by `option(amap_key = 'key')`
+#' # or set by key argument in `getAdmin()`
 #'
 #' #Get subordinate administrative regions as a XML
 #' getAdmin('Sichuan Province', output = 'XML') %>%