+++ NEWS.md
@@ -1,3 +1,10 @@
+# amapGeocode 0.3.1
+* Improve the examples of functions
+
+# amapGeocode 0.3.0
+* Added ability to batch process requests for `getAdmin()`, `getLocation()`, `getCoord()`, `convertCoord()` functions
+* Tag `convertCoord()` as an Experimental function
+
 # amapGeocode 0.2.0
 * Add other three main functions: `getAdmin()`, `extractAdmin()` and `convertCoord()`